Chore: Update routing strategy
