+++ src/routers/root.tsx
@@ -1,33 +1,34 @@
 import { createBrowserRouter } from 'react-router-dom';
 import LoginPage from '../pages/login';
-import UsersPage from '../pages/users';
 import UserManagement from '../pages/users/management';
 import UsersList from '../pages/users/list';
 import ErrorPage from '../pages/error-page';
-import { getLoginToken } from '../utils/auth';
+import SpecificUserPage from '../pages/users/user';
+import Router from '../pages';
 
 export const router = createBrowserRouter([
-  {
-    path: '/',
-    element: getLoginToken() ? <UsersPage /> : <LoginPage />,
-    errorElement: <ErrorPage />,
-  },
   {
     path: '/login',
     element: <LoginPage />,
   },
   {
-    path: '/users',
-    element: <UsersPage />,
+    path: '/',
+    element: <Router />,
     errorElement: <ErrorPage />,
     children: [
       {
-        path: '/users',
+        path: 'users',
         element: <UsersList />,
-      },
-      {
-        path: 'management',
-        element: <UserManagement />,
+        children: [
+          {
+            path: 'management',
+            element: <UserManagement />,
+          },
+          {
+            path: 'user/:id',
+            element: <SpecificUserPage />,
+          },
+        ],
       },
     ],
   },
